Reverts previous commit and implements the actual fix
Previous commit was the wrong solution. The issue that arose was because of sub-texel size textures having their read length improperly calculated to intra-texel boundaries, resulting in corruption. It has been fixed by removing the bpps metric and instead directly using the format texel information to calculate this values rather than a lot of arithmetic (that was also wrong).